+++ projects/igniteui-angular/directives/src/directives/notification/notifications.directive.ts
@@ -49,7 +49,7 @@ export abstract class IgxNotificationsDirective extends IgxToggleDirective
      *  `container` is an instance of `HTMLElement`.
      */
     @Input()
-    public container: HTMLElement;
+    public useContainer: boolean;
 
 
     /**
@@ -100,7 +100,6 @@ export abstract class IgxNotificationsDirective extends IgxToggleDirective
             closeOnEscape: false,
             closeOnOutsideClick: false,
             modal: false,
-            target: this.container,
             outlet: this.outlet
         };