+++ src/app/overlay/overlay.sample.ts
@@ -1,4 +1,4 @@
-import { Component, ViewChild, ElementRef, ChangeDetectorRef, OnInit } from '@angular/core';
+import { Component, ViewChild, ElementRef, ChangeDetectorRef, OnInit, viewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import {
     IgxDropDownComponent,
@@ -21,7 +21,8 @@ import {
     IgxButtonDirective,
     IgxRippleDirective,
     IgxDropDownItemComponent,
-    IChangeCheckboxEventArgs
+    IChangeCheckboxEventArgs,
+    IgxToggleDirective
 } from 'igniteui-angular';
 import { IAnimationParams } from 'igniteui-angular/animations';
 
@@ -29,7 +30,7 @@ import { IAnimationParams } from 'igniteui-angular/animations';
     selector: 'overlay-sample',
     styleUrls: ['overlay.sample.css'],
     templateUrl: './overlay.sample.html',
-    imports: [IgxRadioComponent, FormsModule, IgxSwitchComponent, IgxInputGroupComponent, IgxInputDirective, IgxLabelDirective, IgxButtonDirective, IgxRippleDirective, IgxDragDirective, IgxDropDownComponent, IgxDropDownItemComponent]
+    imports: [IgxRadioComponent, FormsModule, IgxSwitchComponent, IgxInputGroupComponent, IgxInputDirective, IgxLabelDirective, IgxButtonDirective, IgxRippleDirective, IgxDragDirective, IgxDropDownComponent, IgxDropDownItemComponent, IgxToggleDirective]
 })
 export class OverlaySampleComponent implements OnInit {
     @ViewChild(IgxDropDownComponent, { static: true })
@@ -38,8 +39,10 @@ export class OverlaySampleComponent implements OnInit {
     private button: ElementRef;
     @ViewChild(IgxDragDirective, { static: true })
     private igxDrag: IgxDragDirective;
-    @ViewChild('outlet', { static: true })
-    private outlet: ElementRef;
+    @ViewChild('container', { static: true })
+    private container: ElementRef;
+    @ViewChild('containerTarget', { static: true, read: IgxToggleDirective })
+    private containerTarget: IgxToggleDirective;
 
     public items = [];
     public itemsCount = 10;
@@ -58,9 +61,9 @@ export class OverlaySampleComponent implements OnInit {
     public scrollStrategy = 'NoOp';
     public closeOnOutsideClick = true;
     public modal = true;
-    public useOutlet = false;
+    public useContainer = false;
     public hasAnimation = true;
-    public changeOutlet = false;
+    public changeContainer = false;
     public animationLength = 300; // in ms
 
     private xAddition = 0;
@@ -137,7 +140,7 @@ export class OverlaySampleComponent implements OnInit {
                         this.verticalStartPoint = 'Middle';
                         this.closeOnOutsideClick = true;
                         this.modal = true;
-                        this.useOutlet = true;
+                        this.useContainer = true;
                         document.getElementById('mcd').classList.add('selected');
                         document.getElementById('mcsp').classList.add('selected');
                         break;
@@ -226,7 +229,7 @@ export class OverlaySampleComponent implements OnInit {
             positionStrategy: stringMapping['PositionStrategy'][this.positionStrategy],
             scrollStrategy: stringMapping['ScrollStrategy'][this.scrollStrategy],
             modal: this.modal,
-            closeOnOutsideClick: this.closeOnOutsideClick,
+            closeOnOutsideClick: this.closeOnOutsideClick
         };
         this._overlaySettings.positionStrategy.settings.verticalDirection =
             stringMapping['VerticalDirection'][this.verticalDirection];
@@ -236,10 +239,6 @@ export class OverlaySampleComponent implements OnInit {
             stringMapping['HorizontalDirection'][this.horizontalDirection];
         this._overlaySettings.positionStrategy.settings.horizontalStartPoint =
             stringMapping['HorizontalStartPoint'][this.horizontalStartPoint];
-        if (this.useOutlet) {
-            this._overlaySettings.outlet = this.outlet.nativeElement;
-            this._overlaySettings.target = null;
-        }
     }
 
     public onSwitchChange(ev: IChangeCheckboxEventArgs) {
@@ -250,24 +249,23 @@ export class OverlaySampleComponent implements OnInit {
             case 'modal':
                 this._overlaySettings.modal = ev.checked;
                 break;
-            case 'outlet':
-                this._overlaySettings.outlet = ev.checked ? this.outlet.nativeElement : null;
+            case 'container':
                 break;
-            case 'changeOutlet':
+            case 'changeContainer':
                 if (ev.checked) {
-                    this.outlet.nativeElement.style.position = 'fixed';
-                    this.outlet.nativeElement.style.width = '600px';
-                    this.outlet.nativeElement.style.height = '400px';
-                    this.outlet.nativeElement.style.border = '1px solid red';
-                    this.outlet.nativeElement.style.top = '50px';
-                    this.outlet.nativeElement.style.left = '50px';
+                    this.container.nativeElement.style.position = 'fixed';
+                    this.container.nativeElement.style.width = '600px';
+                    this.container.nativeElement.style.height = '400px';
+                    this.container.nativeElement.style.border = '1px solid red';
+                    this.container.nativeElement.style.top = '50px';
+                    this.container.nativeElement.style.left = '50px';
                 } else {
-                    this.outlet.nativeElement.style.position = 'static';
-                    this.outlet.nativeElement.style.width = 'unset';
-                    this.outlet.nativeElement.style.height = 'unset';
-                    this.outlet.nativeElement.style.border = '0';
-                    this.outlet.nativeElement.style.top = 'unset';
-                    this.outlet.nativeElement.style.left = 'unset';
+                    this.container.nativeElement.style.position = 'static';
+                    this.container.nativeElement.style.width = 'unset';
+                    this.container.nativeElement.style.height = 'unset';
+                    this.container.nativeElement.style.border = '0';
+                    this.container.nativeElement.style.top = 'unset';
+                    this.container.nativeElement.style.left = 'unset';
                 }
                 break;
         }
@@ -366,19 +364,17 @@ export class OverlaySampleComponent implements OnInit {
         e.target.classList.add('selected');
     }
 
-    public toggleDropDown() {
+    public toggle() {
         if (this.igxDropDown.collapsed) {
-            this.items = [];
-            for (let item = 0; item < this.itemsCount; item++) {
-                this.items.push(`Item ${item}`);
+            if (this.positionStrategy !== 'Container' && this.positionStrategy !== 'Global') {
+                this.items = [];
+                for (let item = 0; item < this.itemsCount; item++) {
+                    this.items.push(`Item ${item}`);
+                }
             }
             this.cdr.detectChanges();
             this.onChange2();
             this._overlaySettings.target = this.button.nativeElement;
-            if (this.useOutlet) {
-                this._overlaySettings.outlet = this.outlet.nativeElement;
-                this._overlaySettings.target = null;
-            }
             (this._overlaySettings.positionStrategy.settings.openAnimation.options.params as IAnimationParams).duration
                 = `${this.animationLength}ms`;
             (this._overlaySettings.positionStrategy.settings.closeAnimation.options.params as IAnimationParams).duration
@@ -388,7 +384,11 @@ export class OverlaySampleComponent implements OnInit {
                 this._overlaySettings.positionStrategy.settings.closeAnimation = null;
             }
         }
-        this.igxDropDown.toggle(this._overlaySettings);
+        if (this.positionStrategy === 'Container' || this.positionStrategy === 'Global') {
+            this.containerTarget.toggle(this._overlaySettings);
+        } else {
+            this.igxDropDown.toggle(this._overlaySettings);
+        }
     }
 
     public ngOnInit(): void {
@@ -422,7 +422,7 @@ export class OverlaySampleComponent implements OnInit {
         }
     }
 
-    protected moveContainer(target: HTMLElement, direction: string, distance: number): void{
+    protected moveHost(target: HTMLElement, direction: string, distance: number): void{
         const currentTop = parseInt(target.style.top, 10) || 0;
         const currentLeft = parseInt(target.style.left, 10) || 0;